Make the session cookie Path attribute configurable
This was unset, leaving it up to the user agent to decide what the
effective path should be (e.g., parent of the path that responded with
Set-Cookie).

While this is fine for something simple like a SPA served from an
ancestor path of the API it interacts with (e.g., / and /api/<ver>),
having a bit more control over the choice of attribute is desirable.