Simplify and expand SessionManager tests
This change:
* Drops the use of the (mini)Redis-backed store in SessionManager tests,
  in favor of a simple stub implementation.
* Makes use of the new stub to expand coverage for store error cases.
* Makes minor commentary or naming consistency tweaks to existing test
  cases.

In addition, TODOs are added relating to an oversight in the way Clear
is currently structured: if Del succeeds but Set fails, we could be left
in the situation where the user agent is referencing a non-existent
session. This is fine in practice, as it will be replaced with a
pre-session on the next request. However, let's get it right.