Reorder operations in SessionManager.Clear
Since the critical operation in Clear is the allocation of a new
pre-session, the call to Create now comes first and determines overall
success of the Clear call.

The SessionStore.Del call for the previous session is no considered
non-critical, and unexpected errors are now simply logged.

See also discussion in commit message for cbe83e0.