Make GetSession to a free function and rename Get
Previously, GetSession was a SessionManager method simply for the
convenience of not having to not specify the session data playload type
D (i.e., it does not access any field members of SessionManager).

This is awkward in cases where a managed handler has access to the
request context, but cannot access the underlying SessionManager.

In addition, session.GetSession is an awkwardly repetitive name.

Together, GetSession is now Get (i.e., session.Get).