Imbue tokens with a version identifier
The idea here is to allow forward-compatible evolution of the token
implementation over time.

A token version acts as an identifier for that implementation, which
includes token structure and authentication mechanism (e.g., MAC).

Authenticator is now a simple wrapper over the v0 token impl, which has
been lifted out.

The particular form of the authenticated segment of the token (i.e.
"v0!<base64url payload>") should be free of canonicalization issues.